better xlst meta character cleaning
We grew frustrated that metacharacters would leave empty newline after report generation so we made some change to the xlst generation function so that :
It leaves less unclosed tag (there shouldn't be any really but we never know)
It shouldn't leave any empty paragraph tag behind. This has the big advantage that the metacharaters don't leave any unwanted newline after the report has been generated.

We also documented which statements to be alone on a newline. This should be more or less the same than with the previous function, but users might have to sightly adjust their report. Thus, this pull request should be 

¬ finding:::DREAD_SCORE > 1 ¬ µzzzµ needs to be alone on a newline
end of condition (¥) needs to be on a newline
end of for-each (∆) needs to be on a newline
† DREAD_SCORE > 1 †' needs to be on a newline
ƒcodeƒ OUTPUT needs to be on a newline

kudos to 4B3l0 for diging into the dark art of word XML